Add files via upload
rearranged the menu into a more logical order and removed the rest of the unnecessary couts in the menu, but otherwise it seemed pretty clean to begin with. Exception occurs in common.h at line 118 when I run this. This didn't happen when I first downloaded all the most up-to-date code, but now it still happens even if I change everything back to exactly the way it was after I downloaded it. So I have no idea what is going on.