Combined Misc Patches
Pulled the code from the 3 new patches on github and combined them with the TestBed code to make sure they all fit together. Implimented new code to output room schedules with classes listed by the days they meet. Fixed a bug where deleting the course array threw an error if the array was not filled while running the program.